RW-2716 begin react conversion of list-modifier-page component
